Return non-existing role if the .security index is not found (elastic/x-pack#3895)
We used to be very lenient with all kinds of exceptions related to the
`.security` index. Yet, sometimes in tests the index is not yet there but
transport clients already pinging the node this causes issues and transport
clients disconnect. Now if the index is not present we simply return no role.

Original commit: elastic/x-pack@044276bd283ddd0e6db209018f8e66fe78175262