unknown run as users should not be allowed to execute any APIs
If a authenticated user with run as permission attempts to run as an unknown user, the unknown
user will be assigned the default role and anonymous role if enabled. This change prevents this
from happening as we require the run as user to have been looked up by a realm.

Closes elastic/x-pack#3878

Original commit: elastic/x-pack@5fec898828e5732b79531565ed7f2a63fd6d9a64