Moved over to using ES Settings infra (elastic/x-pack#204)
This change removes PrelertSettings in favour of using the ES settings infrestructure. This means that all settings that used to go in engine_api.yml now goes in the elasticsearch.yml file instead. The settings name have not yet been renamed to be prefixed with `_xpack.prelert` this will be done later when coordinated with the changes to the function tests that will be needed for it

Closes elastic/x-pack#201
Closes elastic/x-pack#202

Original commit: elastic/x-pack@b98c595e7d8950585f5cf02e2b20ff98e03cb4f8