Get C++ processes to log through a named pipe to the JVM (elastic/x-pack#243)
This means the C++ processes no longer need to log errors to stderr
as well as log4cxx. Instead the most recent 5 errors are returned to the
REST endpoint if the process crashes.

For this initial version the C++ log messages are put into the main
Elasticsearch log file. In future this could be changed to put them in
an index instead or as well.

Original commit: elastic/x-pack@dc3630ade49594be221ef7bb3e79bfaa3f280817