Make HttpRequest toXContent / parse method coherent (elastic/x-pack#4153)
This commit changes the toXContent() method so that it generates content that can be parsed again by the parse() method. Before this commit, the authentication of the HTTP request is rendered as:

{ "auth": {"username": "foo", "password": "bar" } }

but the parsing method expects the authentication type to be a root object:

{ "auth":  { "basic" :  {"username": "foo", "password": "bar" } } }

The toXContent method has been adapted to include the type of authentication in the generated content.

(cherry picked from commit elastic/x-pack@954c875)

Original commit: elastic/x-pack@3190b9ea17889e57eb9c9496052f59ce76f02a8e