Watcher: Throw exception if HttpClient response is not a HTTP response (elastic/x-pack#4154)
If the HTTP response is an invalid one, it is still logged as success.
This commit changes the behaviour, that if the response status code is
set to -1 (which means it could not be interpreted), than an IOException
is thrown and thus the execution will be marked as a failure.

Closes elastic/x-pack#4152

Original commit: elastic/x-pack@a54dfe035d3771b02ea6f3b2b336adc1cf404a66