security: improve migrate tool output and remove trappy config option
This commit improves the output of the migrate tool in cases when there are errors parsing entries
in the roles or users files. This is done through the use of a logger that delegates its output to
the terminal.

Additionally, the `-c` option has been removed. This option was used to set the configuration directory
but this should be handled one way only and that is through the use of the `-Epath.conf` setting.

Closes elastic/x-pack#3757
Closes elastic/x-pack#3758

Original commit: elastic/x-pack@3a4e7a581dd61a92b5ca73f5b685e23be5842c2f