Watcher: Compile scripts on each invocation (elastic/x-pack#4239)
Transform and condition scripts were only compiled on its initial creation, so
when a new watch is created or when the master node loads all the watches. However
changing a script (like a stored one) did not lead to any changes in the in memory
watch store and thus the old script was executed again.

We do however have a mechanism in Elasticsearch's ScriptService that already does some
caching, and should reuse that one.

Closes elastic/x-pack#4237

Original commit: elastic/x-pack@7b6544389aa3747805b6d677652bb68a030fe400