Build: Convert xplugins to use new extra projects setup (elastic/x-pack#4175)
* Build: Convert xplugins to use new extra projects setup

This change makes the gradle initialization for xplugins look in the
correct location for elasticsearch, which is now as a sibling of an
elasticsearch-extra directory, with x-plugins as a child of the extra
directory.

The elasticsearch side of this change is
elastic/elasticsearchelastic/x-pack#21773. This change will enable renaming x-plugins
to x-pack, see elastic/x-pack#3643.

Original commit: elastic/x-pack@17ed0ebe7ce36d14539068693efcbc6bc4f4bbb8