Watcher: Add deprecation logging for unencoded URLs (elastic/x-pack#4442)
Watcher: Add deprecation logging for unencoded URLs

With the switch to apache HTTP client in 6.0 all URLs require
proper URL encoding. This is a BWC breaking change and thus we
need to inform the user about this upfront so he has the chance
to change his configured URLs.

The first idea was to do the check on start when all the watches
are loaded, however as URLs can templatable and can change with each
execution, this was not feasible.

This implementation remembers the last URL for each http input and webhook
action and only logs the the current URL has changed. The main reason for this
is to not flood the depreciation log with the same URL for a the same watch
over and over again and just log if new URLs are encountered.

Relates elastic/x-pack#4434

Original commit: elastic/x-pack@c672a82818de3e4ae8018095e118ce8ea4095800