Watcher: Add dedicated acknowledged state (elastic/x-pack#4588)
In order to display a better UI page for the watch history, the state
of a throttled watch should be visualized. However, right now there
is no way to differentiate between a time throttled watch and a user
acknowledged watch (or action). This commit introduces a new type in a
throttled result, which in turn is used to set the execution state of
a watch.

Closes elastic/x-pack#4531

Original commit: elastic/x-pack@0fe144852aa8a2d147b7fc2cd2561821614e4aef