Don't clear realm-cache during security upgrade on basic license (elastic/x-pack#4806)
If a user has a basic license, but previously had a full-featured license (e.g. a trial license that expired) then they may have .security index that needs to be migrated even though security is not allowed under their license.

This change makes the clearing of the realm-cache conditional on the license state. If X-Pack is running on a license that does not allow auth, then, when the `logstash_system` user is disabled as part of an upgrade migration, the cache is not cleared.

This change also fix a bug whereby a mapping update could take place even if a data migration was in progress, which could cause the `logstash_system` user to be temporarily enabled when it ought not be.

(cherry picked from commit elastic/x-pack@cd37d0726ba9e4cb898a23ba9001f681226ed37e)

Original commit: elastic/x-pack@a292fcbd1edd8970eb02a0085df779994588ddf6