Watcher: Include watch status in history (elastic/x-pack#4875)
In order to help watcher UI easily gather the status of a watch
and its actions, we should write the watch status in the history
as well.

This commit is doing exactly that. Tests ensure, that the status is not searchable,
as in the watch mapping itself.

This also requires the history template to be changed, thus the counter of the template
had to be increased as well.

Some minor refactorings by making ctors private that dont need to be public
have been happening as well.

Closes elastic/x-pack#4735

Original commit: elastic/x-pack@29814a48987765ca49fb1a12f5c8126ecaef1df1