stop using random generated keys for signing and don't sign when TLS is enabled (#547)
This commit changes the logic for the crypto service signing text so that the randomly generated key
is only used when sending data to older nodes. When sending data to nodes of the same version or
higher the text will be signed if TLS is disabled, a system key is present, and signing is enabled.

The changes made here will allow us to move closer to removing the CryptoService class usage for sending
data on the transport layer.