[Security] Treat LDAP exceptions as auth failures (#565)
Any failure to communicate with an LDAP server should be treated as if authentication failed rather than a fatal "stop the world" error.

Because the AD realm validates the user's password by attempting to bind to the directory with that username/password combination, connection failures are to be expected and must be handled gracefully because the AD realm may not be the last realm in the order-chain, and any exception would prevent later realms from being tested.

(cherry picked from commit 5282290b0090cd1484713f3ba0cd8161bf34b262)