Test: remove logstash_system from old and mixed cluster tests
The logstash_system user does not exist in the version we test against as the old version (currently 5.0) so these
tests always fail. We can still test this user in the upgraded cluster since we know it will exist after an upgrade and
there are things we should verify such as the user being disabled on a upgraded cluster.