Fix settings deprecations tests
This commit fixes some tests for settings deprecations. The previous
tests relied on the fact that the warning headers only test for
containment rather than exact match. With the change that made this test
strict, a convenience method for testing settings deprecations was
added. This commit migrates the existing settings deprecation tests to
use this convenience method.