Watcher: Fix watch serialization after starting/stopping (#666)
Whenever watcher is started/stopped/the master node changes, it
reloads all the watches from disk. With the 5.x release we introduced
a possibility to fall back to the old scripting language (groovy), if
none was specified.

However this mechanism also kicks in, when the reloading happens, even
though the watch was created under 5.x - this could effectively lead
to a scripting language set to groovy, even though it was never intended
to be set.

This commit adds a check before loading the watches to find out, when
the watch index had been created and only uses the BWC compatible
mechanism, if the index was created before 5.0.

Note: This does not eliminate the problem completely, but reduces it
only.

Closes #542